Renamed rigger files
