+++ gt/tools/auto_rigger/rig_module_arm_biped.py
@@ -4,7 +4,7 @@ github.com/TrevisanGMW/gt-tools
 """
 from gt.tools.auto_rigger.rigger_utils import RiggerConstants, find_objects_with_attr, find_proxy_node_from_uuid
 from gt.utils.attr_utils import hide_lock_default_attrs, set_attr_state, set_attr
-from gt.tools.auto_rigger.rigger_framework import Proxy, ModuleGeneric
+from gt.tools.auto_rigger.rig_framework import Proxy, ModuleGeneric
 from gt.tools.auto_rigger.rigger_utils import get_proxy_offset
 from gt.utils.transform_utils import match_translate, Vector3
 from gt.utils.naming_utils import NamingConstants
@@ -285,7 +285,7 @@ if __name__ == "__main__":
     logger.setLevel(logging.DEBUG)
     cmds.file(new=True, force=True)
 
-    from gt.tools.auto_rigger.rigger_framework import RigProject
+    from gt.tools.auto_rigger.rig_framework import RigProject
     a_arm = ModuleBipedArm()
     a_arm_rt = ModuleBipedArmRight()
     a_arm_lf = ModuleBipedArmLeft()
