+++ gt/tools/auto_rigger/rig_module_digit_biped.py
@@ -2,10 +2,10 @@
 Auto Rigger Digit Modules (Fingers, Toes)
 github.com/TrevisanGMW/gt-tools
 """
-from gt.tools.auto_rigger.rigger_utils import RiggerConstants, find_objects_with_attr, find_proxy_node_from_uuid
+from gt.tools.auto_rigger.rig_utils import RiggerConstants, find_objects_with_attr, find_proxy_node_from_uuid
 from gt.utils.attr_utils import hide_lock_default_attrs, set_attr_state, set_attr
-from gt.tools.auto_rigger.rigger_framework import Proxy, ModuleGeneric
-from gt.tools.auto_rigger.rigger_utils import get_proxy_offset
+from gt.tools.auto_rigger.rig_framework import Proxy, ModuleGeneric
+from gt.tools.auto_rigger.rig_utils import get_proxy_offset
 from gt.utils.transform_utils import match_translate, Vector3
 from gt.utils.naming_utils import NamingConstants
 from gt.utils.color_utils import ColorConstants
@@ -506,7 +506,7 @@ if __name__ == "__main__":
     logger.setLevel(logging.DEBUG)
     cmds.file(new=True, force=True)
 
-    from gt.tools.auto_rigger.rigger_framework import RigProject
+    from gt.tools.auto_rigger.rig_framework import RigProject
     a_digit_mod = ModuleBipedDigits()
     a_digit_mod_lf = ModuleBipedFingersLeft()
     a_digit_mod_rt = ModuleBipedFingersRight()
