+++ gt/tools/auto_rigger/rig_module_leg_biped.py
@@ -2,11 +2,11 @@
 Auto Rigger Leg Modules
 github.com/TrevisanGMW/gt-tools
 """
-from gt.tools.auto_rigger.rigger_utils import RiggerConstants, find_objects_with_attr, find_proxy_node_from_uuid
+from gt.tools.auto_rigger.rig_utils import RiggerConstants, find_objects_with_attr, find_proxy_node_from_uuid
 from gt.utils.attr_utils import add_attr, hide_lock_default_attrs, set_attr_state, set_attr
-from gt.tools.auto_rigger.rigger_framework import Proxy, ModuleGeneric
-from gt.tools.auto_rigger.rigger_utils import get_proxy_offset
+from gt.tools.auto_rigger.rig_framework import Proxy, ModuleGeneric
 from gt.utils.transform_utils import match_translate, Vector3
+from gt.tools.auto_rigger.rig_utils import get_proxy_offset
 from gt.utils.naming_utils import NamingConstants
 from gt.utils.color_utils import ColorConstants
 from gt.utils.curve_utils import get_curve
@@ -358,7 +358,7 @@ if __name__ == "__main__":
     logger.setLevel(logging.DEBUG)
     cmds.file(new=True, force=True)
 
-    from gt.tools.auto_rigger.rigger_framework import RigProject
+    from gt.tools.auto_rigger.rig_framework import RigProject
     a_proxy = Proxy()
     a_proxy.set_initial_position(y=84.5)
     a_proxy.set_name("test")
