+++ gt/tools/auto_rigger/rig_modules.py
@@ -1,0 +1,25 @@
+from gt.tools.auto_rigger.rig_framework import ModuleGeneric
+from gt.tools.auto_rigger.rig_module_leg_biped import (ModuleBipedLeg,
+                                                       ModuleBipedLegLeft,
+                                                       ModuleBipedLegRight)
+from gt.tools.auto_rigger.rig_module_spine import ModuleSpine
+from gt.tools.auto_rigger.rig_module_arm_biped import (ModuleBipedArm,
+                                                       ModuleBipedArmLeft,
+                                                       ModuleBipedArmRight)
+from gt.tools.auto_rigger.rig_module_digit_biped import (ModuleBipedDigits,
+                                                         ModuleBipedFingersLeft,
+                                                         ModuleBipedFingersRight)
+
+
+class RigModules:
+    ModuleGeneric = ModuleGeneric
+    ModuleBipedLeg = ModuleBipedLeg
+    ModuleBipedLegLeft = ModuleBipedLegLeft
+    ModuleBipedLegRight = ModuleBipedLegRight
+    ModuleSpine = ModuleSpine
+    ModuleBipedArm = ModuleBipedArm
+    ModuleBipedArmLeft = ModuleBipedArmLeft
+    ModuleBipedArmRight = ModuleBipedArmRight
+    ModuleBipedDigits = ModuleBipedDigits
+    ModuleBipedFingersLeft = ModuleBipedFingersLeft
+    ModuleBipedFingersRight = ModuleBipedFingersRight
