--- gt/tools/auto_rigger/rigger_modules.py
@@ -1,25 +1,0 @@
-from gt.tools.auto_rigger.rigger_framework import ModuleGeneric
-from gt.tools.auto_rigger.rigger_module_leg_biped import (ModuleBipedLeg,
-                                                          ModuleBipedLegLeft,
-                                                          ModuleBipedLegRight)
-from gt.tools.auto_rigger.rigger_module_spine import ModuleSpine
-from gt.tools.auto_rigger.rigger_module_arm_biped import (ModuleBipedArm,
-                                                          ModuleBipedArmLeft,
-                                                          ModuleBipedArmRight)
-from gt.tools.auto_rigger.rigger_module_digit_biped import (ModuleBipedDigits,
-                                                            ModuleBipedFingersLeft,
-                                                            ModuleBipedFingersRight)
-
-
-class RigModules:
-    ModuleGeneric = ModuleGeneric
-    ModuleBipedLeg = ModuleBipedLeg
-    ModuleBipedLegLeft = ModuleBipedLegLeft
-    ModuleBipedLegRight = ModuleBipedLegRight
-    ModuleSpine = ModuleSpine
-    ModuleBipedArm = ModuleBipedArm
-    ModuleBipedArmLeft = ModuleBipedArmLeft
-    ModuleBipedArmRight = ModuleBipedArmRight
-    ModuleBipedDigits = ModuleBipedDigits
-    ModuleBipedFingersLeft = ModuleBipedFingersLeft
-    ModuleBipedFingersRight = ModuleBipedFingersRight
