+++ gt/tools/auto_rigger/template_biped.py
